- add styles to sign in page - split sign in form component
